uint64

+++ state/indexer/sink/psql/psql.go
@@ -70,8 +70,8 @@ func runInTransaction(db *sql.DB, query func(*sql.Tx) error) error {
 // queryWithID executes the specified SQL query with the given arguments,
 // expecting a single-row, single-column result containing an ID. If the query
 // succeeds, the ID from the result is returned.
-func queryWithID(tx *sql.Tx, query string, args ...interface{}) (uint32, error) {
-	var id uint32
+func queryWithID(tx *sql.Tx, query string, args ...interface{}) (uint64, error) {
+	var id uint64
 	if err := tx.QueryRow(query, args...).Scan(&id); err != nil {
 		return 0, err
 	}
@@ -83,7 +83,7 @@ func queryWithID(tx *sql.Tx, query string, args ...interface{}) (uint32, error) 
 //
 // If txID > 0, the event is attributed to the Tendermint transaction with that
 // ID; otherwise it is recorded as a block event.
-func insertEvents(dbtx *sql.Tx, blockID, txID uint32, evts []abci.Event) error {
+func insertEvents(dbtx *sql.Tx, blockID, txID uint64, evts []abci.Event) error {
 	// Populate the transaction ID field iff one is defined (> 0).
 	var txIDArg interface{}
 	if txID > 0 {
